Reattach videooutput on too many failures
I verified that sometimes AVPlayerItemVideoOutput may become stuck due to
broken media, strange network behaviours, or when media ends and the
user skips back to a valid position. This can result in
- no video rendering
- crashes (xcode sees the device as disconnected, no stack trace)

When this happens the video output needs to be recreated and reattached
ref. https://stackoverflow.com/questions/24800742/iosavplayeritemvideooutput-hasnewpixelbufferforitemtime-doesnt-work-correctly